requirements.txt
fastapi==0.116.0
joblib==1.5.1
numpy==2.3.1
pandas==2.3.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
scikit-learn==1.7.0
scipy==1.16.0
uvicorn==0.35.0
xgboost==3.0.2
